Update README.md
## Title
#### Chingu Solo-Project: Tickybot Clone
## #Overview
This is my first project. And the main goal for working on it has been to fulfill the requirement by Chingu to join the development team named voyage 34. Despite this being the main goal, during the development process, I have been able to learn many things. Some of the things I have learned are some Knowledge gaps that I have started to work on. So one can say that the overall purpose of the project is practice.